+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## 0.1.0 Under development
 
+- Enh #2: Add configuration for the widget to `FormModelInterface::class` (@terabytesoftw)
 - Initial commit